(feat) : move the landing page to the pages folder
